update export-components command and move helper components into different directory
