+++ package.json
@@ -6,13 +6,13 @@
     "type": "git",
     "url": "git+https://github.com/department-of-veterans-affairs/design-system.git"
   },
-  "main": "dist/js/components",
+  "main": "dist/js",
   "scripts": {
+    "export-components": "rimraf dist/js && babel src/components --out-dir dist/js/components --ignore .spec.jsx,.njk,.yml && babel src/helpers --out-dir dist/js/helpers",
     "prestart": "node prebuild.js",
-    "export-components": "babel src/components --out-dir dist/js/components --ignore .spec.jsx,.njk,.yml",
     "start": "fractal watch",
-    "test": "karma start --single-run",
     "site": "fractal build-site && gh-pages -d build",
+    "test": "karma start --single-run",
     "test-watch": "karma start --auto-watch",
     "lint:js": "eslint --quiet --ext .js --ext .jsx .",
     "test-debug": "karma start --browsers Chrome"
@@ -66,6 +66,7 @@
     "react-test-renderer": "^15.5.0",
     "recursive-readdir": "^2.2.1",
     "resolve-url-loader": "^2.2.0",
+    "rimraf": "^2.6.2",
     "sass-loader": "^6.0.6",
     "sinon": "^4.1.6",
     "style-loader": "^0.19.0",
