+++ src/components/errorable_text_input/ErrorableTextInput.jsx
@@ -4,7 +4,7 @@ import _ from 'lodash';
 
 import ToolTip from '../tooltip/ToolTip.jsx';
 
-import { makeField } from '../../model/fields.js';
+import { makeField } from '../../helpers/fields.js';
 
 /**
  * A form input with a label that can display error messages.
